+++ dolphinscheduler-task-plugin/dolphinscheduler-task-spark/src/main/java/org/apache/dolphinscheduler/plugin/task/spark/SparkVersion.java
@@ -18,48 +18,5 @@
 package org.apache.dolphinscheduler.plugin.task.spark;
 
 public enum SparkVersion {
-
-    /**
-     * 0 SPARK1
-     * 1 SPARK2
-     * 2 SPARK1SQL
-     * 3 SPARK2SQL
-     */
-    SPARK1(0, "SPARK1", "${SPARK_HOME1}/bin/spark-submit", "SPARK1"),
-    SPARK2(1, "SPARK2", "${SPARK_HOME2}/bin/spark-submit", "SPARK2"),
-
-    SPARK1SQL(2, "SPARK1SQL", "${SPARK_HOME1}/bin/spark-sql", "SPARK1"),
-
-    SPARK2SQL(3, "SPARK2SQL", "${SPARK_HOME2}/bin/spark-sql", "SPARK2");
-
-    private final int code;
-    private final String descp;
-    /**
-     * usage: spark-submit [options] <app jar | python file> [app arguments]
-     */
-    private final String command;
-    private final String sparkVersion;
-
-    SparkVersion(int code, String descp, String command, String sparkVersion) {
-        this.code = code;
-        this.descp = descp;
-        this.command = command;
-        this.sparkVersion = sparkVersion;
-    }
-
-    public int getCode() {
-        return code;
-    }
-
-    public String getDescp() {
-        return descp;
-    }
-
-    public String getCommand() {
-        return command;
-    }
-
-    public String getSparkVersion() {
-        return sparkVersion;
-    }
+    SPARK1, SPARK2
 }
